Moved all the various promises custom types in utils/promises
